feat(header/header.html.org): remove inclusions of css/js in html generation
fix the problem in #29

The problem caused by including 'jquery.js' twice inside of each HTML
page for a chapter and outside of it (index.html). This fixes the
problem.

As a side effect, individual HTML file do not work as it did before. For
example, '01_Introduction.html' will not render correctly without being
embedded inside of 'index.html'.