Update DependentTypeTheory.lean
I am being pedantic here but the following sentence is technically incorrect as it is written:

"Like sections, nested namespaces have to be closed in the order they are opened"

The intended meaning is that they should be closed in the reverse order. I suggest an alternative wording here, but perhaps another alternative could be saying "last-in, first-out" more directly.